Change config.json format and fix SensorArduino code
