+++ src/arduino/sensor_arduino/PressureSensor.cpp
@@ -1,18 +1,18 @@
-#include <PressureSensor.hpp>
+#include "PressureSensor.hpp"
 
 #define MAX_PRESSURE 1000
 #define ROOM_PRESSURE 15
 
-float PressureSensor::mapVal(float val, float lower1, float upper1, float lower2, float upper2) {
+PressureSensor::PressureSensor(int pin) {
+    this->pin = pin;
+    pinMode(pin, OUTPUT);
+}
+
+float mapVal(float val, float lower1, float upper1, float lower2, float upper2) {
     float diff1 = upper1 - lower1;
     float diff2 = upper2 - lower2;
     float factor = diff2 / diff1;
     return (val - lower1) * factor + lower2;
-}
-
-PressureSensor::PressureSensor(int pin) {
-    this->pressurePin = pin;
-    pinMode(pressurePin, OUTPUT);
 }
 
 float PressureSensor::getPressure() {
