+++ src/arduino/sensor_arduino/PressureSensor.hpp
@@ -1,14 +1,11 @@
-#include <Arduino.h>
-
 #ifndef PRESSURE_SENSOR_HPP
 #define PRESSURE_SENSOR_HPP
 
 class PressureSensor {
     private:
-        float mapVal(float val, float lower1, float upper1, float lower2, float upper2);
-        int pressurePin;
 
     public:
+        int pin;
         PressureSensor(int pin);
         float getPressure();
 };
