+++ src/arduino/sensor_arduino/SensorArduino.hpp
@@ -1,10 +1,10 @@
-#include <Thermocouple.hpp>
-#include <PressureSensor.hpp>
-#include <map>
-#include <Wire.h>
-#include <chrono>
-#include <thread>
+#include "Arduino.h"
+#include "Thermocouple.hpp"
+#include "PressureSensor.hpp"
+#include <vector>
+#include <string>
 #include <cstdint>
+#include <Wire>
 
 #ifndef SENSOR_ARDUINO_HPP
 #define SENSOR_ARDUINO_HPP
@@ -13,21 +13,21 @@ class SensorArduino {
   private:
     // these maps map the pin to its respective sensor
 
-    std::map<int, Thermocouple> thermocouples;
-    std::map<int, PressureSensor> pressure_sensors;
+    std::vector<Thermocouple> thermocouples;
+    std::vector<PressureSensor> pressure_sensors;
 
     // send data to pi every 50 milliseconds
     const int SEND_DELAY = 50;
 
-    float getThermo();
-    float getPressure();
     void receiveData();
     void registerSensors();
+    int recvI2CByte();
 
   public:
     SensorArduino::SensorArduino();
     void read();
     void sendData();
+    void error();
 };
 
 #endif
