+++ src/arduino/sensor_arduino/Thermocouple.hpp
@@ -1,4 +1,3 @@
-#include <utility>
 #include <vector>
 #include <Adafruit_MAX31856.h>
 
@@ -9,6 +8,7 @@ class Thermocouple {
     private:
         Adafruit_MAX31856 *maxthermo;
     public:
+        vector<int> pins;
         Thermocouple::Thermocouple(vector<int> pins);
         Thermocouple::~Thermocouple();
         std::pair<float, bool> getTemp();
