+++ src/arduino/sensor_arduino/sensor_arduino.ino
@@ -1,4 +1,4 @@
-#include <SensorArduino.hpp>
+#include "SensorArduino.hpp"
 
 SensorArduino sensor_arduino;
