+++ src/pi/notes.txt
@@ -1,0 +1,1 @@
+FOR ARDUINO UNO, A0 = 14. THIS ISN'T THE SAME FOR OTHER ARDUINO MODELS. THIS NEEDS TO BE CHANGED IN THE CONFIG.JSON
